requirements.txt
Django==5.0.6
django-crispy-forms==2.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
Jinja2==3.1.3
phonenumbers==8.13.40
psycopg2-binary==2.9.9
PyJWT==2.8.0